:technologist: Distribute own fixtures as pytest plugin
So that downstream packages/projects can also make use of these
test helpers rather than having to duplicate them.
