+++ simple_certmanager/test/plugin.py
@@ -1,0 +1,8 @@
+"""
+Pytest plugin.
+"""
+
+from .fixtures import encrypted_keypair  # noqa: F401
+from .fixtures import leaf_keypair  # noqa: F401
+from .fixtures import root_cert  # noqa: F401
+from .fixtures import root_key  # noqa: F401
